put Ring struct in a separate file
